Rename dyn to ref
