+++ mlir/lib/Conversion/MQTOptToMQTRef/CMakeLists.txt
@@ -6,5 +6,6 @@
 #
 # Licensed under the MIT License
 
-add_subdirectory(MQTRefToMQTOpt)
-add_subdirectory(MQTOptToMQTRef)
+file(GLOB CONVERSION_SOURCES *.cpp)
+
+add_mlir_library(MQTOptToMQTRef ${CONVERSION_SOURCES} DEPENDS MQTOptToMQTRefIncGen)
