+++ mlir/lib/Dialect/CMakeLists.txt
@@ -7,5 +7,5 @@
 # Licensed under the MIT License
 
 add_subdirectory(MQTOpt)
-add_subdirectory(MQTDyn)
+add_subdirectory(MQTRef)
 add_subdirectory(MQTMem)
