--- mlir/lib/Dialect/MQTDyn/IR/MQTDynOps.cpp
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2023 - 2025 Chair for Design Automation, TUM
- * Copyright (c) 2025 Munich Quantum Software Company GmbH
- * All rights reserved.
- *
- * SPDX-License-Identifier: MIT
- *
- * Licensed under the MIT License
- */
-
-#include "mlir/Dialect/MQTDyn/IR/MQTDynDialect.h"
-
-// The following headers are needed for some template instantiations.
-// IWYU pragma: begin_keep
-#include <llvm/ADT/TypeSwitch.h>
-#include <mlir/IR/Builders.h>
-#include <mlir/IR/DialectImplementation.h>
-// IWYU pragma: end_keep
-
-//===----------------------------------------------------------------------===//
-// Dialect
-//===----------------------------------------------------------------------===//
-
-#include "mlir/Dialect/MQTDyn/IR/MQTDynOpsDialect.cpp.inc"
-
-void mqt::ir::dyn::MQTDynDialect::initialize() {
-  addTypes<
-#define GET_TYPEDEF_LIST
-#include "mlir/Dialect/MQTDyn/IR/MQTDynOpsTypes.cpp.inc"
-      >();
-
-  addOperations<
-#define GET_OP_LIST
-#include "mlir/Dialect/MQTDyn/IR/MQTDynOps.cpp.inc"
-      >();
-}
-
-//===----------------------------------------------------------------------===//
-// Types
-//===----------------------------------------------------------------------===//
-
-#define GET_TYPEDEF_CLASSES
-#include "mlir/Dialect/MQTDyn/IR/MQTDynOpsTypes.cpp.inc"
-
-//===----------------------------------------------------------------------===//
-// Interfaces
-//===----------------------------------------------------------------------===//
-
-#include "mlir/Dialect/MQTDyn/IR/MQTDynInterfaces.cpp.inc"
-
-//===----------------------------------------------------------------------===//
-// Operations
-//===----------------------------------------------------------------------===//
-
-#define GET_OP_CLASSES
-#include "mlir/Dialect/MQTDyn/IR/MQTDynOps.cpp.inc"
