+++ mlir/lib/Dialect/MQTRef/IR/MQTRefOps.cpp
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2023 - 2025 Chair for Design Automation, TUM
+ * Copyright (c) 2025 Munich Quantum Software Company GmbH
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
+#include "mlir/Dialect/MQTRef/IR/MQTRefDialect.h"
+
+// The following headers are needed for some template instantiations.
+// IWYU pragma: begin_keep
+#include <llvm/ADT/TypeSwitch.h>
+#include <mlir/IR/Builders.h>
+#include <mlir/IR/DialectImplementation.h>
+// IWYU pragma: end_keep
+
+//===----------------------------------------------------------------------===//
+// Dialect
+//===----------------------------------------------------------------------===//
+
+#include "mlir/Dialect/MQTRef/IR/MQTRefOpsDialect.cpp.inc"
+
+void mqt::ir::ref::MQTRefDialect::initialize() {
+  addTypes<
+#define GET_TYPEDEF_LIST
+#include "mlir/Dialect/MQTRef/IR/MQTRefOpsTypes.cpp.inc"
+      >();
+
+  addOperations<
+#define GET_OP_LIST
+#include "mlir/Dialect/MQTRef/IR/MQTRefOps.cpp.inc"
+      >();
+}
+
+//===----------------------------------------------------------------------===//
+// Types
+//===----------------------------------------------------------------------===//
+
+#define GET_TYPEDEF_CLASSES
+#include "mlir/Dialect/MQTRef/IR/MQTRefOpsTypes.cpp.inc"
+
+//===----------------------------------------------------------------------===//
+// Interfaces
+//===----------------------------------------------------------------------===//
+
+#include "mlir/Dialect/MQTRef/IR/MQTRefInterfaces.cpp.inc"
+
+//===----------------------------------------------------------------------===//
+// Operations
+//===----------------------------------------------------------------------===//
+
+#define GET_OP_CLASSES
+#include "mlir/Dialect/MQTRef/IR/MQTRefOps.cpp.inc"
